requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
duckdb==0.8.0
airflow-provider-duckdb==0.1.0
astro-sdk-python[duckdb]==1.6.1
pandas==2.0.2